requirements.txt
streamlit==1.38.0
Pillow==10.4.0
# Optional high-quality upscaler (fallback to Lanczos if missing)
realesrgan==0.3.0
torch
torchvision
torchaudio
